search implemented with all orders display also
